test: Extract runtime Info to new package
This will allow Go SDK users to access this previously internal package.
